Add files via upload
Time has been split up along with what day of the week the class is on. DotW for classes is a simple boolean value (yes or no). 

!!VERY IMPORTANT!!
DO NOT RUN THE datainput.py if the database has already been created. I have no idea why but it will re-add the course values even though I told it to ignore them.