Add files via upload
To make things a bit easier, this is the "main" file that I'll be using for the documentation. Once you other fellas finish your code you can update it to here and that's what I'll use to finish the documentation. For now, it's just a copy of the "myfirst.py" under Kevin's folder.